+++ data-plane/core/src/main/java/dev/knative/eventing/kafka/broker/core/ReactiveKafkaConsumer.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package dev.knative.eventing.kafka.broker.dispatcher;
+package dev.knative.eventing.kafka.broker.core;
 
 import io.vertx.core.Future;
 import io.vertx.core.Handler;
