+++ data-plane/core/src/main/java/dev/knative/eventing/kafka/broker/core/ReactiveKafkaProducer.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package dev.knative.eventing.kafka.broker.receiver;
+package dev.knative.eventing.kafka.broker.core;
 
 import io.vertx.core.Future;
 import org.apache.kafka.clients.producer.Producer;
