+++ data-plane/dispatcher/src/test/java/dev/knative/eventing/kafka/broker/dispatcher/MockReactiveKafkaConsumer.java
@@ -15,6 +15,7 @@
  */
 package dev.knative.eventing.kafka.broker.dispatcher;
 
+import dev.knative.eventing.kafka.broker.core.ReactiveKafkaConsumer;
 import io.vertx.core.Future;
 import io.vertx.core.Handler;
 import io.vertx.core.Promise;
