+++ data-plane/receiver-vertx/src/test/java/dev/knative/eventing/kafka/broker/receiververtx/impl/ReceiverVerticleTracingVertxImplTest.java
@@ -15,7 +15,7 @@
  */
 package dev.knative.eventing.kafka.broker.receiververtx.impl;
 
-import dev.knative.eventing.kafka.broker.receiver.ReactiveProducerFactory;
+import dev.knative.eventing.kafka.broker.core.ReactiveProducerFactory;
 import dev.knative.eventing.kafka.broker.receiver.impl.ReceiverVerticleTracingTest;
 import dev.knative.eventing.kafka.broker.receiververtx.VertxProducerFactory;
 import io.cloudevents.CloudEvent;
